+++ payme-sdk/src/main/java/vn/payme/sdk/model/DataMethod.kt
@@ -1,10 +1,10 @@
 package vn.payme.sdk.model
 
 class DataMethod {
-    var linkedId : Double? = null
+    var linkedId : Long? = null
     var swiftCode : String? = null
     var issuer : String? = null
-    constructor(linkedId : Double?,swiftCode : String?,issuer:String?){
+    constructor(linkedId : Long?,swiftCode : String?,issuer:String?){
         this.linkedId = linkedId
         this.swiftCode = swiftCode
         this.issuer = issuer
